+++ renderer-compose-multiplatform/public/src/androidInstrumentedTest/kotlin/software/amazon/app/platform/presenter/molecule/backgesture/ForwardBackPressEventsToPresentersComposeTest.kt
@@ -10,7 +10,7 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.test.assertTextEquals
-import androidx.compose.ui.test.junit4.AndroidComposeTestRule
+import androidx.compose.ui.test.junit4.v2.AndroidComposeTestRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.test.espresso.Espresso
 import androidx.test.ext.junit.rules.ActivityScenarioRule
@@ -28,7 +28,9 @@ class ForwardBackPressEventsToPresentersComposeTest {
 
   @get:Rule val activityRule = ActivityScenarioRule(TestActivity::class.java)
 
-  @get:Rule val composeTestRule = AndroidComposeTestRule(activityRule, ::getActivityFromTestRule)
+  @get:Rule
+  val composeTestRule =
+    AndroidComposeTestRule(activityRule, activityProvider = { getActivityFromTestRule(it) })
 
   @Test
   fun back_press_events_are_forwarded_to_presenters() {