+++ renderer-compose-multiplatform/public/src/desktopTest/kotlin/software/amazon/app/platform/renderer/ComposeRendererTest.kt
@@ -10,7 +10,7 @@ import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.test.ExperimentalTestApi
 import androidx.compose.ui.test.assertTextEquals
 import androidx.compose.ui.test.onNodeWithTag
-import androidx.compose.ui.test.runComposeUiTest
+import androidx.compose.ui.test.v2.runComposeUiTest
 import kotlin.test.Test
 import kotlinx.coroutines.flow.MutableStateFlow
 import software.amazon.app.platform.presenter.BaseModel