+++ robot-compose-multiplatform/public/src/appleAndDesktopTest/kotlin/software/amazon/app/platform/robot/ComposeRobotTest.kt
@@ -8,7 +8,7 @@ import androidx.compose.ui.test.ExperimentalTestApi
 import androidx.compose.ui.test.SemanticsNodeInteractionsProvider
 import androidx.compose.ui.test.assertTextEquals
 import androidx.compose.ui.test.onNodeWithTag
-import androidx.compose.ui.test.runComposeUiTest
+import androidx.compose.ui.test.v2.runComposeUiTest
 import assertk.assertFailure
 import assertk.assertThat
 import assertk.assertions.isFalse